+++ projects/angular-auth-oidc-client/src/lib/callback/refresh-session.service.ts
@@ -43,7 +43,10 @@ export class RefreshSessionService {
             );
         }
 
-        return forkJoin([this.startRefreshSession(), this.silentRenewService.refreshSessionWithIFrameCompleted$.pipe(take(1))]).pipe(
+        return forkJoin([
+            this.startRefreshSession(customParams),
+            this.silentRenewService.refreshSessionWithIFrameCompleted$.pipe(take(1)),
+        ]).pipe(
             timeout(this.configurationProvider.openIDConfiguration.silentRenewTimeoutInSeconds * 1000),
             retryWhen(this.timeoutRetryStrategy.bind(this)),
             map(([_, callbackContext]) => {