+++ projects/angular-auth-oidc-client/src/lib/config/default-config.ts
@@ -32,5 +32,5 @@ export const DEFAULT_CONFIG: OpenIdConfiguration = {
     customParams: {},
     eagerLoadAuthWellKnownEndpoints: true,
     disableRefreshIdTokenAuthTimeValidation: false,
-    tokenRefreshInSeconds: 3,
+    tokenRefreshInSeconds: 4,
 };