+++ projects/schematics/src/ng-add/actions/configs.ts
@@ -50,6 +50,16 @@ const AZURE_AD_REFRESH_TOKENS = `{
             autoUserinfo: false,
     }`;
 
+const AUTH_0 = `{
+            stsServer: '<stsUrlOrTenantId>',
+            redirectUrl: window.location.origin,
+            clientId: 'please-enter-auth0-clientId',
+            scope: 'openid profile offline_access',
+            responseType: 'code',
+            silentRenew: true,
+            useRefreshToken: true,
+        }`;
+
 const OIDC_PLAIN = `{
             stsServer: '<stsUrlOrTenantId>',
             redirectUrl: window.location.origin,
@@ -61,4 +71,4 @@ const OIDC_PLAIN = `{
             renewTimeBeforeTokenExpiresInSeconds: 10,
         }`;
 
-export { DEFAULT_CONFIG, AZURE_AD_SILENT_RENEW, IFRAME_SILENT_RENEW, AZURE_AD_REFRESH_TOKENS, OIDC_PLAIN };
+export { DEFAULT_CONFIG, AZURE_AD_SILENT_RENEW, IFRAME_SILENT_RENEW, AZURE_AD_REFRESH_TOKENS, OIDC_PLAIN, AUTH_0 };